+++ Modules/Intent.Modules.AspNetCore.Controllers.Dispatch.ServiceContract/FactoryExtensions/ServiceContractControllerInstaller.cs
@@ -119,6 +119,14 @@ namespace Intent.Modules.AspNetCore.Controllers.Dispatch.ServiceContract.Factory
                         if (!operationModel.InternalElement.HasStereotype("Synchronous"))
                         {
                             awaitModifier = "await ";
+                        }
+
+                        // if NOT async or IS async but the "no cancellation token" property is not true
+                        if (!operationModel.InternalElement.HasStereotype("Synchronous") &&
+                            !(operationModel.InternalElement.HasStereotype("Asynchronous") &&
+                             operationModel.InternalElement.GetStereotype("Asynchronous").TryGetProperty("2801e2a9-5797-406f-b289-43af8fbb2d7e", out var property) &&
+                             property.Value == "true"))
+                        {
                             arguments = string.IsNullOrEmpty(arguments)
                                 ? "cancellationToken"
                                 : $"{arguments}, cancellationToken";