+++ Modules/Intent.Modules.Application.ServiceCallHandlers/release-notes.md
@@ -1,3 +1,7 @@
+### Version 4.4.1
+
+- Improvement: Prevent `CancellationToken` automatically applied with `Asyncronous` stereotype applied with `No Cancellation Token` set to true.
+
 ### Version 4.4.0
 
 - Improvement: Updated dependency modules.