+++ Modules/Intent.Modules.Aws.Lambda.Functions.Dispatch.Services/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.3
+
+- Improvement: Prevent `CancellationToken` automatically applied with `Asyncronous` stereotype applied with `No Cancellation Token` set to true.
+
 ### Version 1.0.2
 
 - Improvement: Updated to use `IMessageBus` instead of `IEventBus` (variable name changed from `eventBus` to `messageBus`). Maintains backward compatibility by supporting both `IEventBus` and `IMessageBus` interfaces.