+++ Modules/Intent.Modules.FastEndpoints.Dispatch.MediatR/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.1
+
+- Improvement: Prevent `CancellationToken` automatically applied with `Asyncronous` stereotype applied with `No Cancellation Token` set to true.
+
 ### Version 1.0.0
 
 - New Feature: Adds MediatR Dispatch support for the `Intent.FastEndpoints` module.